Final Project/Final Project/requirements.txt
asgiref==3.7.2
beautifulsoup4==4.12.2
certifi==2023.7.22
charset-normalizer==3.2.0
cssselect==1.2.0
Django==4.2.4
django-ckeditor==6.7.0
django-environ==0.11.2
django-jazzmin==2.6.0
django-js-asset==2.1.0
django-ranged-response==0.2.0
django-recaptcha==3.0.0
idna==3.4
lxml==4.9.3
markdown2==2.4.10
Pillow==10.0.0
pyquery==2.0.0
readtime==3.0.0
requests==2.31.0
soupsieve==2.5
sqlparse==0.4.4
typing_extensions==4.7.1
urllib3==2.0.4
whitenoise==6.5.0
